Added caching for default UserDao annotations from spring boot cache. The user model is now cached when loading and saving methods are invoked. When the save method is invoked, the old cache is cleared. Added a return value for the save method to correctly cache the new user model after saving.
Added abstract class CacheCleaner with it interface BaseCacheCleaner and implemented it in UserDaoImpl for cleaning cache.

Added throwing EntitySavingException when the save method in UserDao is malfunctioning.

Caching is enabled using the @EnableCaching annotation in CacheConfig.

Added logging to the methods for loading and saving the UserServiceImpl class.

Added caching UserInfo after loading and caching UserInfo after saving tests in UserDaoCacheTest. Added auxiliary config classes UserEntityTestConfig and UserTestConfig