During the mapping of the UserInfo object to UserEntity, UserEntity.id was populated with null due to the absence of this field in UserInfo. Since only UserInfo objects were being cached, retrieving the complete UserEntity object required querying the database every time.
This led to other issues, such as the inability to delete a user from the database using the UserDao.delete() method. The deletion was performed based on the username field, searching for this field in the cache, finding a UserInfo object, incompletely mapping it to UserEntity, failing to locate the full object in the database, and consequently, not deleting it.

This problem was addressed by changing the cache storage system: now UserEntity objects were stored in the cache instead of UserInfo objects (where the id field was absent). With this modification, accessing the cache allowed obtaining the complete UserEntity object directly.

The method for deleting a user in UserRepository was also modified from delete() to deleteUserEntityByUsername(). Deletion based on the username proved to be more efficient.

The approach to removing UserEntity from the cache was altered as well: instead of using the @CacheEvict annotation, deletion now occurs through the Cache.evict() method. Using the annotation required first retrieving the complete object from the cache or database, deleting it, and then returning it from the UserDao.delete() method. In contrast, using Cache.evict() only requires passing the username as the key to this method.