Added @MappingWrapper annotation generation code. This annotation is applied to an interface that needs to automatically convert its parameters or return types for use in the 'source' class (this class has the same methods but different parameters or return types). The MappingWrapper processor then generates the implementation of the annotated interface, wraps the 'source' class, transforms all necessary values using the 'mapper', generates methods with the same names as those of the 'source' class, and calls the methods of the nested class inside them.
Added MappingWrapperClassCodeGenerator for generating code for a MappingWrapper class that implements a class marked with the @MappingWrapper annotation. This class utilizes the MappingWrapperClassCodeBuilder for step-by-step construction of the MappingWrapper code text (using the new MappingWrapperMethodCodeBuilder).

Added MappingWrapperAnnotationService and its implementation, which extract details such as 'source' and 'mapper' parameters from the @MappingWrapper annotation, such as MappingWrapperAnnotationDetails, MappingWrapperAnnotationParamFullNames, MappingWrapperNameDetails. Providing these values is necessary to eliminate unnecessary dependencies of the MappingWrapperClassCodeGenerator on the 'Element' class.

Added ElementMethodService and its implementation, which extract method signature from the 'Element' class. Providing this value is necessary to eliminate unnecessary dependencies of the MappingWrapperClassCodeGenerator on the 'Element' class.