Added MainContextTest in the mapping-wrapper module to extend the Spring context to other tests. Also added 'MainMappingWrapperConfig' for injecting the Spring context into the project without a SpringBootApplication class in this module.
Added MappingWrapperClassCodeServiceTest to test class code generation. This test method generates the implementation of the MappingWrapper interface with mappings or unchanged parameters, data return types, as well as with void and null parameters.
Also added testing with an incorrect source class where there are methods that differ from those present in the interface.
Also added testing with an incorrect mapper class where there are insufficient methods for converting the required values.