Improved the method of constructing the code for the MappingWrapper class in the MappingWrapperClassCodeBuilderImpl. Previously, all class elements were assembled in a StringBuilder object, and now they are assembled in a specialized class called MappingWrapperClassCode, which includes fields, methods, etc. This class now only contains the class data without unnecessary empty lines and other components.
Added the MappingWrapperClassParser class, which transforms the test code text of the class into a MappingWrapperClassCode object.

Introduced the MappingWrapperCodeStringBuilder class, responsible for converting the MappingWrapperClassCode object into a string.

Changed the code generation test system for the MappingWrapper class: previously, the generated class code texts were compared with the code from the mapping_wrapper_class.example file. Now, we take the code text from the file, transform it into a MappingWrapperClassCode object, and systematically check if each code component matches the expected result.

Added the mono_mapping_wrapper_class.example file, containing the expected result of generating code text for the MappingWrapper class with generics. This is necessary to test the correctness of the generated code with generics. As an example, Mono<> is used.

Test instances of the interface, source code, and mapper for the ClassDetails class have been moved to separate configuration files: InterfaceClassDetailsTestConfig, SourceClassDetailsTestConfig, MapperClassDetailsTestConfig.

Changed the name of the CachedUserDao class to CachedUserEntityDao. Also, the return values of the UserDao class have been changed from UserEntity to UserInfo.