Resolved a bug related to UPDATING UserInfo.
Previously, updating UserInfo in the database was unsuccessful due to the id field from UserEntity remaining null during the conversion process. This led to JPA attempting to add a new UserEntity instead of updating the existing one and creating a new id for it.

Removed the update() method from CachedUserEntityDao and introduced a new class, UpdatableUserEntityDao, which includes this method. In this class, the UserEntity is loaded by username, the id is extracted from it, and then added to the updating UserEntity. This modified model is then passed to the UserRepository for updating.