Established a one-to-many reference of MoodTag in MoodEntity. Established a many-to-one reference of UserEntity in MoodTagEntity.
Implemented a conversion method in UserEntityMapper for transforming moodEntryEntities to moodTagsEntities.

Renamed mergeMoodTagsWithSameNameInEntities() to addDbReferences() in CachedUserEntityDao.

Integrated the addDbReferences() method invocation in the update method of UpdatableUserEntityDao, making it transactional.
Removed nullable ID checking from the addDbReferences() method in UpdatableUserEntityDao class, as all entities are required to retrieve all database references due to the absence of IDs in the UserInfo business-logic model.

Encapsulated the functionality of the load() method in UserServiceImpl within Mono, ensuring that any exceptions thrown are caught by the controller.