Added the moodTags field to the UserInfo class to pass tags that are currently not associated with any mood states.
Renamed the CachedMoodTagEntityLoaderDao interface to MoodTagEntityLoaderDao.

Removed the exception thrown in case the loadByUserId() method in the CachedMoodTagEntityLoaderDaoImpl class returned an empty list.

Added loading of mood states in the returned UserInfoDto class within the default range of days. The date range loading parameter can be explicitly configured via the application.yml configuration file.

Made the save() and update() methods of the MoodEntryServiceImpl class transactional by annotating them with @Transactional, as it first merges tags and then adds a MoodEntry.

Added a null check when converting collections in mappers to avoid NullPointerExceptions.

Added error logging in the update() method of the MoodEntryServiceImpl class.