Fixed a bug: if an exception was thrown during the addition or modification of a MoodEntry, the MoodTags associated with that MoodEntry were still being added to the database.
Merged MoodTags and the addition/modification of MoodEntry into a single class: MoodEntryWithTagsDao. Marked the methods saveEntryWithTags() and updateEntryWithTags() with @Transactional to combine them into a transaction. Now, if tags are added but adding the mood state throws an exception, the database will roll back along with the new tags.