Added user registration capability through the signup() method of UserSignupController in the authorization-service microservice.
This method receives a request with username and rowPassword fields, encrypts the password, and passes the user's DTO object to the database-service microservice, which in turn saves it in the database.

Access to the user saving method of the database-service microservice is provided through the UserDatabaseConnector class located in the domain module. It is based on the HttpClient utility, simplifying access to WebClient from the webflux library.

Since UserDatabaseConnector in the domain module requires dependencies on layer-connector and logger, to avoid cyclic dependencies, the reference to domain in these modules was removed. Conversely, the necessary class YamlPropertySourcesPlaceholderConfigurerFactory for the correct import of yml variables was added to them.