Fixed a bug where the GET request to api/service/database/mood_tags, intended to retrieve mood tags, returned NO_CONTENT when a request was sent with a non-existent user ID. This is incorrect behavior; it should return a NOT_FOUND status instead.
To address this, a checkUserExistence() check was added to each method in MoodTagService. This method queries the cache or database for a userId and, if a UserInfo object is returned, it means the user exists, and the search for the tag can be attempted.

For this purpose, a loadById() method was added to UserService to enable such a search. Additionally, to avoid querying the database every time to find user data, another cache storage was created, which uses the user id as the key and the UserEntity as the value.

However, the addition of a reference to UserService in MoodTagService resulted in a circular dependency between beans, as MoodTagService already had a reference to UserService. Therefore, methods for loading MoodTags and UserInfo were moved to separate classes: MoodTagLoader and UserLoader, respectively. Also, the loading of MoodTagLoader in UserLoaderConfig was made @Lazy to avoid the circular dependency.

But that was not the last problem encountered. Since the loadById() method used full UserInfo loading with MoodTags and MoodEntries, an infinite call occurred when loading UserInfo in MoodTagService. To resolve this, user loading was split into two methods: loadAll() and load(). In the load() method, MoodTags are now not loaded into UserInfo, preventing an entry into the infinite loop.

Исправилен баг: GET запрос api/service/database/mood_tags, цель которого получение тэгов настроения, возвращал NO_CONTENT, когда к нему посылался запрос с несущетсвующим id пользователя. Это некорректное поведение, должен возвращаться статус NOT_FOUND.

Для этого в была добавлена проверка checkUserExistence() в каждом методе MoodTagService. Данный метод делает запрос в кэш или базу данных по userId и если возвращается объект UserInfo, значит он существует и значит тэг можно попытаться найти.

Для эого был добавлен метод loadById() в UserService для того, чтобы такой поиск мог быть осушествлен.
Так же, чтобы не осуществлять каждый раз запрос в базу данных для поиска данных пользователя, было создано еще одно кэш хранилише, которое имеет ключ: user id, а значение: UserEntity.

Однако из-за добавления ссылки на UserService в MoodTagService, получилась циклическая зависимость бинов. Так как MoodTagService уже содержал ссылку на UserService. Поэтому методы для загрузки MoodTags и UserInfo были вынесены в отдельные классы: MoodTagLoader и UserLoader соответсвенно. А так же загрузка MoodTagLoader в UserLoaderConfig сделана @Lazy для избежания циклической зависимости.

Но это не последняя проблема, с которой пришлось столкнуться. Так как в методе loadById() использовалась загрузка полного UserInfo с MoodTags и MoodEntries, поэтому возникал бесконечный вызов при закгрузке UserInfo в MoodTagService. Для этого загрузка пользователя была разделена на 2 метода: loadAll() и load(). В методе load() в UserInfo теперь не загружаются MoodTags, поэтому входа в бесконечный цикл не происходит.