Fixed bug with missing user data caching.
Removed caching of user existence. Instead, added UserEntityExistenceDao class, which loads the user from cache or database and checks for null.

Fixed bug with missing user data caching. Access to the classes responsible for caching through cache spring annotations was done not through the object, but through super. Because of this, AOP did not intercept the method call, and the caching annotations did not trigger. The problem was fixed by replacing the class relationship of UserEntityDao from inheritance to composition. Added UserEntityDaoFacade facade to combine the functionality of UserEntityUpdaterDao, UserEntityExistenceDao, UserEntityLoaderDao.

In the MoodTagService class, the checkUserExistence() method was edited. Now it does not load the user and check for null, but directly calls the userExistenceService.existsById() method.

Удалено кэширование существования пользователя. Вместо этого добавлен класс UserEntityExistenceDao, который загружает из кэша или базы данных пользователя и делает проверку на null.

Исправлен баг с отсуствием кэширования пользвовательских данных. Обращение к классам, отвечающим за кэширование с помощью аннотаций cache spring, проходило не через объект, а через super. Из-за этого AOP не перехватывало обращение к методу и аннотации кэширования не срабатывали. Проблема исправлена путем замены отношений классов UserEntityDao с наследования на композицию. Добавлен фасад UserEntityDaoFacade для объелинения функциональности UserEntityUpdaterDao, UserEntityExistenceDao, UserEntityLoaderDao.

В MoodTagService классе был заменен метод checkUserExistence(). Теперь он не загружает пользователя и проверяет на null, а сразу обращается к методу userExistenceService.existsById()